Separate constructor from the base alpha router
